+++ cifstr/cifstr.vcxproj
@@ -20,13 +20,19 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\application.cpp" />
-    <ClCompile Include="src\cifexcept.ixx" />
-    <ClCompile Include="src\ciffile.ixx" />
-    <ClCompile Include="src\cifparse.ixx" />
-    <ClCompile Include="src\cifstr.ixx" />
-    <ClCompile Include="src\logger.ixx" />
-    <ClCompile Include="src\pdqciflib.ixx" />
-    <ClCompile Include="src\util.ixx" />
+    <ClCompile Include="src\cifstr.cpp" />
+    <ClCompile Include="src\cifstr.hpp" />
+    <ClCompile Include="src\vendor\row\pdqciflib\cifexcept.cpp" />
+    <ClCompile Include="src\vendor\row\pdqciflib\ciffile.cpp" />
+    <ClCompile Include="src\vendor\row\pdqciflib\cifparse.cpp" />
+    <ClCompile Include="src\vendor\row\pdqciflib\util.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="src\vendor\row\pdqciflib.hpp" />
+    <ClInclude Include="src\vendor\row\pdqciflib\cifexcept.hpp" />
+    <ClInclude Include="src\vendor\row\pdqciflib\ciffile.hpp" />
+    <ClInclude Include="src\vendor\row\pdqciflib\cifparse.hpp" />
+    <ClInclude Include="src\vendor\row\pdqciflib\util.hpp" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <VCProjectVersion>16.0</VCProjectVersion>
