+++ cifstr/cifstr.vcxproj.filters
@@ -18,26 +18,40 @@
     <ClCompile Include="src\application.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="src\cifexcept.ixx">
+    <ClCompile Include="src\cifstr.hpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="src\ciffile.ixx">
+    <ClCompile Include="src\vendor\row\pdqciflib\cifexcept.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="src\cifparse.ixx">
+    <ClCompile Include="src\vendor\row\pdqciflib\util.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="src\cifstr.ixx">
+    <ClCompile Include="src\vendor\row\pdqciflib\ciffile.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="src\logger.ixx">
+    <ClCompile Include="src\vendor\row\pdqciflib\cifparse.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="src\pdqciflib.ixx">
+    <ClCompile Include="src\cifstr.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="src\util.ixx">
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="src\vendor\row\pdqciflib.hpp">
       <Filter>Source Files</Filter>
-    </ClCompile>
+    </ClInclude>
+    <ClInclude Include="src\vendor\row\pdqciflib\cifexcept.hpp">
+      <Filter>Source Files</Filter>
+    </ClInclude>
+    <ClInclude Include="src\vendor\row\pdqciflib\ciffile.hpp">
+      <Filter>Source Files</Filter>
+    </ClInclude>
+    <ClInclude Include="src\vendor\row\pdqciflib\cifparse.hpp">
+      <Filter>Source Files</Filter>
+    </ClInclude>
+    <ClInclude Include="src\vendor\row\pdqciflib\util.hpp">
+      <Filter>Source Files</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
